Breaks up map markers into their own component
